requirements.txt
absl-py==1.4.0
aioice==0.7.6
aiortc==1.4.0
altair==4.2.2
attrs==22.2.0
av==10.0.0
blinker==1.5
cachetools==5.3.0
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
contourpy==1.0.7
cryptography==39.0.1
cycler==0.11.0
decorator==5.1.1
dnspython==2.3.0
entrypoints==0.4
flatbuffers==23.1.21
fonttools==4.38.0
gitdb==4.0.10
GitPython==3.1.31
google-crc32c==1.5.0
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
Jinja2==3.1.2
jsonschema==4.17.3
kiwisolver==1.4.4
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.0
mdurl==0.1.2
mediapipe==0.9.1.0
netifaces==0.11.0
numpy==1.24.2
opencv-contrib-python==4.7.0.72
opencv-python-headless==4.7.0.72
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
protobuf==3.20.3
pyarrow==11.0.0
pycparser==2.21
pydeck==0.8.0
pyee==9.0.4
Pygments==2.14.0
pylibsrtp==0.8.0
Pympler==1.0.1
pyOpenSSL==23.0.0
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
rich==13.3.1
semver==2.13.0
six==1.16.0
smmap==5.0.0
streamlit==1.19.0
streamlit-lottie==0.0.3
streamlit-webrtc==0.44.7
toml==0.10.2
toolz==0.12.0
tornado==6.2
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
validators==0.20.0
zipp==3.15.0
